requirements.txt
watchdog==4.0.1
PyYAML==6.0.1
pathlib==1.0.1